+++ src/main/kotlin/com/github/zero9178/mlirods/language/highlighting/TableGenSemanticTokensAnnotator.kt
@@ -1,10 +1,8 @@
 package com.github.zero9178.mlirods.language.highlighting
 
 import com.github.zero9178.mlirods.color.FIELD
-import com.github.zero9178.mlirods.language.generated.psi.TableGenFieldAccessValue
-import com.github.zero9178.mlirods.language.generated.psi.TableGenFieldBodyItem
-import com.github.zero9178.mlirods.language.generated.psi.TableGenLetBodyItem
-import com.github.zero9178.mlirods.language.generated.psi.TableGenLetItem
+import com.github.zero9178.mlirods.color.PREPROCESSOR_MACRO_NAME
+import com.github.zero9178.mlirods.language.generated.psi.TableGenPreprocessorDirective
 import com.github.zero9178.mlirods.language.psi.TableGenFieldIdentifierNode
 import com.intellij.lang.annotation.AnnotationHolder
 import com.intellij.lang.annotation.Annotator
@@ -19,6 +17,14 @@ private class TableGenSemanticTokensAnnotator : Annotator, DumbAware {
                 holder.newSilentAnnotation(HighlightSeverity.TEXT_ATTRIBUTES)
                     .range(element.fieldIdentifier)
                     .textAttributes(FIELD)
+                    .create()
+            }
+
+            is TableGenPreprocessorDirective -> {
+                val identifier = element.identifier ?: return
+                holder.newSilentAnnotation(HighlightSeverity.TEXT_ATTRIBUTES)
+                    .range(identifier)
+                    .textAttributes(PREPROCESSOR_MACRO_NAME)
                     .create()
             }
         }