+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # MLIR ODS Changelog
 
 ## [Unreleased]
+
+## [0.4.0] - 2025-02-18
 ### Added
 - Added resolution of include statements, enabling the `GoTo` action 
 - Implemented renaming of include statements, enabling the `Rename` action of TableGen files